feat: move all internal libs into separate packages
Moves all internal libraries into separate packages. Ie: trpc-swr -> @trpc-swr/client and
trpc-swr/infinite -> @trpc-swr/infinite

BREAKING CHANGE: trpc-swr is no longer a single packages
