+++ apps/demo/package.json
@@ -15,7 +15,7 @@
 		"@nextui-org/react": "1.0.0-beta.12",
 		"@paralleldrive/cuid2": "^2.2.0",
 		"@radix-ui/react-icons": "^1.3.0",
-		"@trpc-swr/core": "workspace:*",
+		"@trpc-swr/client": "workspace:*",
 		"@trpc/client": "^10.19.1",
 		"@trpc/server": "^10.19.1",
 		"@tw-classed/react": "^1.4.4",
