+++ apps/demo/src/lib/trpc.ts
@@ -1,5 +1,5 @@
 import { AppRouter } from "@/server/_app";
-import { createSWRProxyHooks } from "@trpc-swr/core";
+import { createSWRProxyHooks } from "@trpc-swr/client";
 import { httpBatchLink } from "@trpc/client";
 
 export const api = createSWRProxyHooks<AppRouter>({
