+++ packages/infinite/index.ts
