+++ packages/next/tsconfig.json
@@ -4,26 +4,19 @@
     "esModuleInterop": true,
     "jsx": "react-jsx",
     "lib": ["esnext", "dom"],
-    "module": "NodeNext",
-    "moduleResolution": "nodenext",
+    "moduleResolution": "node",
     "noFallthroughCasesInSwitch": true,
     "noImplicitReturns": false,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "strictBindCallApply": true,
     "outDir": "./dist",
-    "rootDir": "../",
     "strict": true,
     "target": "ES2018",
     "baseUrl": ".",
     "noEmitOnError": true,
-    "paths": {
-      "trpc-swr": ["../core/index.ts"],
-      "trpc-swr/infinite": ["../src/infinite/index.ts"],
-      "trpc-swr/ssg": ["../src/ssg/index.ts"]
-    },
     "incremental": true
   },
-  "types": ["node", "vite/client", "vitest/globals"],
+  "types": ["node"],
   "exclude": ["./**/dist", "examples"]
 }
