+++ packages/ssg/index.ts
