+++ packages/ssg/serialize.ts
