+++ packages/ssg/tsconfig.json
@@ -1,0 +1,22 @@
+{
+  "compilerOptions": {
+    "declaration": true,
+    "esModuleInterop": true,
+    "jsx": "react-jsx",
+    "lib": ["esnext", "dom"],
+    "moduleResolution": "node",
+    "noFallthroughCasesInSwitch": true,
+    "noImplicitReturns": false,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "strictBindCallApply": true,
+    "outDir": "./dist",
+    "strict": true,
+    "target": "ES2018",
+    "baseUrl": ".",
+    "noEmitOnError": true,
+    "incremental": true
+  },
+  "types": ["node"],
+  "exclude": ["./**/dist", "examples"]
+}
