--- packages/trpc-swr/_internal/index.ts
@@ -1,14 +1,0 @@
-export {
-	TRPCContext,
-	type TRPCContextType,
-	useTRPCSWRContext,
-} from "./context";
-export {
-	createSWRHooks,
-	type CreateTRPCSWRHooks,
-	type TRPCProvider,
-	type UseSWR,
-	type UseSWRMutation,
-	getQueryKey,
-} from "./hooks/createSWRHooks";
-export * from "./useTransformedFallback";
