--- packages/trpc-swr/_internal/package.json
@@ -1,17 +1,0 @@
-{
-  "main": "./dist/index.js",
-  "module": "./dist/index.esm.js",
-  "types": "./dist/index.d.ts",
-  "exports": "./dist/index.mjs",
-  "private": true,
-  "scripts": {
-    "watch": "bunchee index.ts -w",
-    "build": "bunchee index.ts"
-  },
-  "peerDependencies": {
-    "react": ">=18.2.0",
-    "@trpc/client": ">=10",
-    "@trpc/server": ">=10",
-    "swr": ">=2"
-  }
-}
