--- packages/trpc-swr/core/index.ts
@@ -1,5 +1,0 @@
-export { createSWRProxyHooks } from "./createTRPCSWR";
-export { createSWRHooks } from "trpc-swr/_internal";
-export type { CreateTRPCSWRHooks } from "trpc-swr/_internal";
-export type { CreateTRPCSWRProxy } from "./createSWRProxyHooks";
-export * from "./types";
