--- packages/trpc-swr/infinite/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "rootDir": "..",
-  },
-  "include": ["."],
-  "exclude": ["./**/dist"]
-}
