--- packages/trpc-swr/package.json
@@ -1,147 +1,0 @@
-{
-  "name": "trpc-swr",
-  "version": "1.0.0-SNAPSHOT.1000",
-  "private": false,
-  "description": "tRPC-ified SWR hooks",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/sachinraja/trpc-swr.git"
-  },
-  "homepage": "https://github.com/sachinraja/trpc-swr#readme",
-  "bugs": {
-    "url": "https://github.com/sachinraja/trpc-swr/issues"
-  },
-  "keywords": [
-    "trpc",
-    "swr",
-    "react",
-    "hooks",
-    "typescript",
-    "api",
-    "rpc",
-    "rest",
-    "http",
-    "fetch",
-    "query",
-    "mutation"
-  ],
-  "author": "Sachin Raja <sachinraja2349@gmail.com>, Sanna Jammeh <hello@sannajammeh.com>",
-  "license": "MIT",
-  "main": "./core/dist/index.js",
-  "module": "./core/dist/index.esm.js",
-  "types": "./core/dist/index.d.ts",
-  "exports": {
-    "./package.json": "./package.json",
-    ".": {
-      "types": "./core/dist/index.d.ts",
-      "import": "./core/dist/index.mjs",
-      "module": "./core/dist/index.esm.js",
-      "require": "./core/dist/index.js"
-    },
-    "./infinite": {
-      "types": "./infinite/dist/index.d.ts",
-      "import": "./infinite/dist/index.mjs",
-      "module": "./infinite/dist/index.esm.js",
-      "require": "./infinite/dist/index.js"
-    },
-    "./next": {
-      "types": "./next/dist/index.d.ts",
-      "import": "./next/dist/index.mjs",
-      "module": "./next/dist/index.esm.js",
-      "require": "./next/dist/index.js"
-    },
-    "./ssg": {
-      "types": "./ssg/dist/index.d.ts",
-      "import": "./ssg/dist/index.mjs",
-      "module": "./ssg/dist/index.esm.js",
-      "require": "./ssg/dist/index.js"
-    },
-    "./_internal": {
-      "types": "./_internal/dist/index.d.ts",
-      "import": "./_internal/dist/index.mjs",
-      "module": "./_internal/dist/index.esm.js",
-      "require": "./_internal/dist/index.js"
-    }
-  },
-  "files": [
-    "core/dist/**/*.{js,d.ts,mjs}",
-    "infinite/dist/**/*.{js,d.ts,mjs}",
-    "next/dist/**/*.{js,d.ts,mjs}",
-    "ssg/dist/**/*.{js,d.ts,mjs}",
-    "_internal/dist/**/*.{js,d.ts,mjs}",
-    "core/package.json",
-    "infinite/package.json",
-    "next/package.json",
-    "ssg/package.json",
-    "_internal/package.json"
-  ],
-  "scripts": {
-    "build": "pnpm build:package _internal && pnpm build:package core && pnpm build:package infinite && pnpm build:package ssg && pnpm build:package next",
-    "watch": "bunchee -w",
-    "dev": "bunchee -w",
-    "build:package": "bunchee index.ts --cwd",
-    "format": "dprint fmt",
-    "lint:format": "dprint check",
-    "lint:types": "tsc",
-    "lint:js": "eslint .",
-    "prepublishOnly": "pnpm build",
-    "test": "vitest run",
-    "test:ui": "vitest --ui"
-  },
-  "peerDependencies": {
-    "@trpc/client": ">=10.18.0",
-    "@trpc/server": ">=10.18.0",
-    "react": ">=17 || >=18",
-    "swr": ">=2"
-  },
-  "devDependencies": {
-    "@faker-js/faker": "^7.6.0",
-    "@sachinraja/eslint-config": "0.2.0",
-    "@testing-library/dom": "9.2.0",
-    "@testing-library/jest-dom": "5.16.5",
-    "@testing-library/react": "14.0.0",
-    "@testing-library/user-event": "14.4.3",
-    "@trpc/client": ">=10.18.0",
-    "@trpc/server": ">=10.18.0",
-    "@types/node": "18.15.11",
-    "@types/react": "18.0.34",
-    "@vitejs/plugin-react-swc": "^3.3.0",
-    "@vitest/ui": "^0.30.0",
-    "bunchee": "3.0.1",
-    "c8": "^7.13.0",
-    "eslint": "8.38.0",
-    "get-port": "6.1.2",
-    "isomorphic-unfetch": "^4.0.2",
-    "jsdom": "21.1.1",
-    "node-fetch": "^3.3.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "rimraf": "^5.0.0",
-    "swr": "2.1.1",
-    "tsup": "^6.7.0",
-    "typescript": "^5.0.4",
-    "undici": "5.21.2",
-    "vite": "4.2.1",
-    "vite-tsconfig-paths": "^4.0.9",
-    "vitest": "0.30.0",
-    "zod": "3.21.4"
-  },
-  "sideEffects": false,
-  "tsup": {
-    "entry": [
-      "src/index.ts",
-      "src/infinite/index.ts",
-      "src/next/index.ts",
-      "src/ssg/index.ts"
-    ],
-    "format": [
-      "esm",
-      "cjs"
-    ],
-    "dts": {
-      "resolve": true
-    },
-    "splitting": true,
-    "clean": true
-  }
-}
