--- packages/trpc-swr/ssg/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "ssg",
-  "main": "./dist/index.js",
-  "module": "./dist/index.esm.js",
-  "types": "./dist/index.d.ts",
-  "exports": "./dist/index.mjs",
-  "private": true,
-  "scripts": {
-    "watch": "bunchee index.ts -w",
-    "build": "bunchee index.ts"
-  },
-  "peerDependencies": {
-    "trpc-swr": "*",
-    "react": ">=18.2.0",
-    "@trpc/client": ">=10",
-    "@trpc/server": ">=10",
-    "swr": ">=2"
-  }
-}
