--- packages/trpc-swr/vite.config.ts
@@ -1,16 +1,0 @@
-/// <reference types="vitest" />
-/// <reference types="vite/client" />
-
-import react from "@vitejs/plugin-react-swc";
-import tsconfigPaths from "vite-tsconfig-paths";
-import { defineConfig } from "vite";
-
-// https://vitejs.dev/config/
-export default defineConfig({
-	plugins: [react(), tsconfigPaths()],
-	test: {
-		globals: true,
-		environment: "jsdom",
-		setupFiles: "./test/setup.ts",
-	},
-});
